Move agent-trace request-end helpers to agents/trace/integration
Per agents/CLAUDE.md, integration helpers belong with request_metrics
and record_llm_metric_tokens. Moves AgentTraceRequestEndState plus the
build/wrap helpers out of preprocessor.rs into agents/trace/integration.
